requirements.txt
pytest==7.4.0
pytest-cov==4.1.0
black==23.3.0
isort==5.12.0
mypy==1.4.1
pre-commit==3.3.3
python-dotenv==1.0.0
ipykernel==6.24.0
jupyter==1.0.0
pydantic==2.1.1
pydantic-settings==2.0.2
langchain==0.0.267
dspy-ai==2.0.0
openai==1.3.0
anthropic==0.5.0
instructor==0.3.1
loguru==0.7.0
typer==0.9.0
rich==13.4.2
tqdm==4.65.0
scipy==1.11.1
numpy==1.25.1
pandas==2.0.3
matplotlib==3.7.2
pymongo==4.4.1
motor==3.2.0
beanie==1.20.0
faiss-cpu==1.7.4
chromadb==0.4.13
pillow==10.0.0
torch==2.0.1
transformers==4.31.0
fastapi==0.100.0
uvicorn==0.23.1
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.1.0
llama-index==0.8.0
haystack-ai==2.0.0
datasets==2.13.1
evaluate==0.4.0
scikit-learn==1.3.0 